Initial cleanup
* Change the codebase to use Cobra for command dispatch.
* Remove the "exp" command, which appears to be WIP that was unfinished.
