+++ .github/workflows/ci.yml
@@ -1,20 +1,21 @@
 name: ci
 on:
   pull_request:
-    branches:
-      - master
   push:
     branches:
       - master
+env:
+  GO_VERSION: '1.18.x'
+
 jobs:
   ci:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
         uses: actions/checkout@v2
-      - name: Install Go 1.16
+      - name: Install ${{env.GO_VERSION}}
         uses: actions/setup-go@v2
         with:
-          go-version: '1.16.x'
+          go-version: ${{env.GO_VERSION}}
       - name: Run Go Build
         run: go build main.go
